chore: move useTeam hook to shared hooks
