+++ apps/trading/lib/hooks/use-team.ts
